Move some non-BlockDev functionality to device.rs
This is stuff that happens before the thing becomes a Stratis BlockDev.
