+++ src/engine/strat_engine/mod.rs
@@ -3,6 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 pub mod blockdev;
+pub mod device;
 pub mod engine;
 pub mod lineardev;
 pub mod metadata;
